Added docker integration
This results in much simpler local running for new onboarded members (or on the webserver)

You can likely further extend this to include production building etc.

I thought i'd make a start on this, but someone else should dive into it to merge it etc.

With these changes you can easily run the backend with the following 2 steps:
1. install docker (desktop)
2. run 'docker-compose up'

start developing :D
it also watches for changes still.